git_sl_copy/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
Flask-Marshmallow 
marshmallow-sqlalchemy
beautifulsoup4 
selenium 
python-dateutil
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Dance==7.1.0
Flask-Login==0.6.3
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
marshmallow==3.23.1
oauthlib==3.2.2
packaging==24.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
URLObject==2.4.3
Werkzeug==3.1.2
WTForms==3.2.1
